Stop storing a VALUE ref in a static variable
Ruby 2.7 introduced GC compaction, which means objects can be
moved around. So unless you register your global variable
to the GC, they might end up invalid.

In this case the variable is never used passed the initialization
so it isn't really a problem, but then we can simply use a local
variable.